Update `--version` to properly handle no network
With no network connection, crashfailing on --version is pretty bad
behavior. Catch the RequestException and just show the current version.
